https://clxmedia.atlassian.net/browse/CLXD-1562
Exporting improvement v2
