+++ clx_budget_management/__manifest__.py
@@ -15,6 +15,7 @@
         'sale_management',
         'sale_subscription',
         'clx_retail_pricelist',
+        'web',
     ],
     'data': [
         'security/ir.model.access.csv',
@@ -24,8 +25,9 @@
         'views/sale_budget_views.xml',
         'views/sale_budget_exports_views.xml',
         'views/res_config_settings_views.xml',
-        # 'views/sale_order_views.xml'
+        'views/tree_view_asset.xml',
     ],
+    "qweb": ["static/src/xml/qweb.xml"],
     'installable': True,
     'application': True,
     'auto_install': False,
