+++ clx_budget_management/static/src/xml/qweb.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="UTF-8"?>
+    <templates id="template" xml:space="preserve">
+
+        <t t-extend="ListView.buttons">
+            <t t-jquery="div.o_list_buttons" t-operation="append">            
+                <t t-if = "widget.modelName == 'sale.budget.exports'">               
+                    <button type="button" class="btn btn-primary generate_exports">
+                        Export Generate
+                    </button>      
+                </t>      
+            </t>
+        </t>
+
+    </templates>
