+++ clx_budget_management/views/tree_view_asset.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+      <data>
+            <template id="assets_backend" name="tree view menu" inherit_id="web.assets_backend">
+                  <xpath expr="." position="inside">
+                        <script type="text/javascript" src="/clx_budget_management/static/src/js/export_tree_view_button.js"></script>
+                  </xpath>
+            </template>
+      </data>
+</odoo>
